renaming App into app
