+++ webapp/src/main.tsx
@@ -1,14 +1,12 @@
+import { TooltipProvider } from '@components/ui/tooltip'
+import QueryClient from "@queries/Client"
+import { QueryClientProvider } from '@tanstack/react-query'
+import 'material-symbols/rounded.css'
 import React from 'react'
 import ReactDOM from 'react-dom/client'
 import { RouterProvider } from 'react-router-dom'
-import { QueryClientProvider } from '@tanstack/react-query'
-
 import './index.css'
-import 'material-symbols/rounded.css'
-
-import QueryClient from "@queries/Client"
 import router from './router'
-import { TooltipProvider } from '@components/ui/tooltip'
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <React.StrictMode>
